requirements.txt
# Core dependencies
torch>=1.12.0
transformers>=4.21.0
datasets>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.64.0

# Training and optimization
deepspeed>=0.8.0
accelerate>=0.18.0
peft>=0.3.0

# Data processing
pyarrow>=8.0.0
tokenizers>=0.13.0

# Utilities
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.5

# Optional: for better performance
flash-attn>=2.0.0; sys_platform == "linux"

# Development dependencies (install with pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# isort>=5.10.0
# mypy>=1.0.0
# pre-commit>=2.20.0

# Documentation dependencies (install with pip install -e ".[docs]")
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0
# myst-parser>=0.18.0